Don't copy __FILE__ string when unnecessary
This reduces the amount of dynamic allocations done before main() runs.
And as a result also the amount of false positives from rudimentary
memory leak detectors (malloc/free counters that use a global's
destructor or atexit() as the moment where all memory should be freed
again).

Additionally this decreases the amount of memory that's detected as
"definitely lost" by Valgrind.