CMake: add libs to targets immediately
Instead of collecting all sources and libraries to link to in several
variables, directly use the respective CMake commands for adding them.

Also introduce an intermediate `INTERFACE` target,
`cucumber-cpp-sources`, to hold everything that's the same for both
`cucumber-cpp-nomain` and `cucumber-cpp`. We might wish to change this
later on to just add everything to `cucumber-cpp-nomain` and make
`cucumber-cpp` link to `cucumber-cpp-nomain`, but that would change the
usage requirements for people using other build systems than CMake.